+++ baseframe/forms/validators.py
@@ -7,6 +7,8 @@ import dns.resolver
 from pyisemail import is_email
 from flask import request
 import wtforms
+from wtforms.validators import (DataRequired, InputRequired, Optional, Length, EqualTo, URL,
+    ValidationError, StopValidation)
 import requests
 from lxml import html
 from coaster.utils import make_name, deobfuscate_email
@@ -15,7 +17,9 @@ from ..signals import exception_catchall
 
 
 __all__ = ['ValidEmail', 'ValidEmailDomain', 'ValidUrl', 'AllUrlsValid', 'StripWhitespace', 'ValidName',
-    'NoObfuscatedEmail', 'ValidCoordinates']
+    'NoObfuscatedEmail', 'ValidCoordinates',
+    # WTForms validators
+    'DataRequired', 'InputRequired', 'Optional', 'Length', 'EqualTo', 'URL', 'ValidationError', 'StopValidation']
 
 
 EMAIL_RE = re.compile(r'\b[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,63}\b', re.I)