+++ CHANGES.rst
@@ -14,7 +14,9 @@
 * Codemirror is now an external dependency installed via Makefile
 * New RadioMatrixField for a table of fields and choices
 * New InlineListWidget to list radio options inline
-
+* Added ParsleyJS validation and reorganized forms to be more modular
+* Added app-global CSRF protection (optional) and removed Mustache dependency
+* Added ``firstline`` filter to extract first line of text from a HTML block
 
 0.3.0
 -----